@@ -18,13 +18,13 @@
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
 		"lint": "prettier --check . && eslint .",
 		"format": "prettier --write .",
-		"format-motoko" : "npx prettier --write --plugin=prettier-plugin-motoko **/*.mo"
+		"format-motoko": "npx prettier --write --plugin=prettier-plugin-motoko **/*.mo"
 	},
 	"devDependencies": {
 		"@rollup/plugin-inject": "^5.0.5",
 		"@sveltejs/adapter-static": "^3.0.1",
-		"@sveltejs/kit": "^2.5.10",
-		"@sveltejs/vite-plugin-svelte": "^3.1.1",
+		"@sveltejs/kit": "^2.15.2",
+		"@sveltejs/vite-plugin-svelte": "^4.0.4",
 		"@types/eslint": "8.56.0",
 		"@types/node": "^20.14.2",
 		"@typescript-eslint/eslint-plugin": "^6.21.0",
@@ -33,30 +33,30 @@
 		"buffer": "^6.0.3",
 		"eslint": "^8.57.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.39.3",
-		"flowbite": "^2.3.0",
-		"flowbite-svelte": "^0.46.1",
-		"flowbite-svelte-icons": "^1.6.1",
+		"eslint-plugin-svelte": "^2.46.1",
+		"flowbite": "^2.5.2",
+		"flowbite-svelte": "^0.47.4",
+		"flowbite-svelte-icons": "^2.0.2",
 		"postcss": "^8.4.38",
 		"prettier": "^3.4.2",
 		"prettier-plugin-motoko": "^0.8.4",
 		"prettier-plugin-organize-imports": "^3.2.4",
-		"prettier-plugin-svelte": "^3.2.4",
-		"svelte": "^4.2.18",
-		"svelte-check": "^3.8.0",
-		"svelte-preprocess": "^5.1.4",
+		"prettier-plugin-svelte": "^3.3.2",
+		"svelte": "^5.17.3",
+		"svelte-check": "^4.1.3",
+		"svelte-preprocess": "^6.0.3",
 		"tailwindcss": "^3.4.4",
 		"tslib": "^2.6.3",
-		"typescript": "^5.4.5",
-		"vite": "^5.2.13"
+		"typescript": "^5.7.3",
+		"vite": "^5.4.11"
 	},
 	"type": "module",
 	"dependencies": {
 		"@dfinity/agent": "^1.4.0",
 		"@dfinity/auth-client": "^1.3.0",
 		"@dfinity/candid": "^1.4.0",
-		"@dfinity/ledger-icrc": "^2.3.3",
+		"@dfinity/ledger-icrc": "^2.7.0",
 		"@dfinity/principal": "^1.4.0",
-		"@dfinity/utils": "^2.3.1"
+		"@dfinity/utils": "^2.8.0"
 	}
 }
