@@ -9,7 +9,7 @@
 	<OptionButton />
 {:else}
 	<button
-		on:click={() => goto('/connect')}
+		onclick={() => goto('/connect')}
 		class="bg-black text-white font-light rounded-full py-4 px-5 text-xs flex flex-row items-center hover:bg-gray-800 active:bg-gray-700 focus:outline-none focus:ring focus:ring-zinc-400"
 	>
 		CONNECT WALLET <div class="pl-1">
