@@ -1,26 +1,35 @@
 <script>
 	// @ts-nocheck
-	export let size = '24';
-	export let color = 'currentColor';
-	export let variation = 'outline';
-	export let ariaLabel = 'bars 3';
-	let viewBox;
-	let svgpath;
+	let {
+		size = '24',
+		color = 'currentColor',
+		variation = 'outline',
+		ariaLabel = 'bars 3',
+		class: className = '',
+		onclick,
+		...restProps
+	} = $props();
+
 	let svgoutline = `<path stroke="${color}" stroke-linecap="round" stroke-linejoin="round" d="M4 6h16M4 12h16M4 18h16"></path> `;
 	let svgsolid = `<path fill="${color}" clip-rule="evenodd" fill-rule="evenodd" d="M3 5a1 1 0 011-1h12a1 1 0 110 2H4a1 1 0 01-1-1zM3 10a1 1 0 011-1h12a1 1 0 110 2H4a1 1 0 01-1-1zM3 15a1 1 0 011-1h12a1 1 0 110 2H4a1 1 0 01-1-1z"></path> `;
-	$: switch (variation) {
-		case 'outline':
-			svgpath = svgoutline;
-			viewBox = '0 0 24 24';
-			break;
-		case 'solid':
-			svgpath = svgsolid;
-			viewBox = '0 0 24 24';
-			break;
-		default:
-			svgpath = svgoutline;
-			viewBox = '0 0 24 24';
-	}
+	let viewBox = $state(svgoutline);
+	let svgpath = $state('0 0 24 24');
+
+	$effect(() => {
+		switch (variation) {
+			case 'outline':
+				svgpath = svgoutline;
+				viewBox = '0 0 24 24';
+				break;
+			case 'solid':
+				svgpath = svgsolid;
+				viewBox = '0 0 24 24';
+				break;
+			default:
+				svgpath = svgoutline;
+				viewBox = '0 0 24 24';
+		}
+	});
 </script>
 
 <svg
@@ -29,13 +38,13 @@
 	tabindex="0"
 	width={size}
 	height={size}
-	class={$$props.class}
-	{...$$restProps}
+	class={className}
+	{...restProps}
 	aria-label={ariaLabel}
 	fill="none"
 	{viewBox}
 	stroke-width="2"
-	on:click
+	{onclick}
 >
 	{@html svgpath}
 </svg>
