@@ -3,20 +3,30 @@
 	import { writable, type Writable } from 'svelte/store';
 	// import ToolbarButton from '../toolbar/ToolbarButton.svelte';
 	import Menu from './Menu.svelte';
-	export let menuClass = 'h-6 w-6 shrink-0';
-	export let onClick = undefined;
-	export let btnClass = 'ms-3';
 	let hiddenStore: Writable<boolean> = getContext('navHidden') ?? writable(true);
 	const toggle = (ev: any) => hiddenStore.update((h) => !h);
 
 	import { ToolbarButton } from 'flowbite-svelte';
+	interface Props {
+		menuClass?: string;
+		onClick?: any;
+		btnClass?: string;
+		[key: string]: any;
+	}
+
+	let {
+		menuClass = 'h-6 w-6 shrink-0',
+		onclick = undefined,
+		btnClass = 'ms-3',
+		...rest
+	}: Props = $props();
 </script>
 
-<ToolbarButton name="Open main menu" on:click={onClick || toggle} {...$$restProps} class={btnClass}>
+<ToolbarButton name="Open main menu" onclick={onclick || toggle} {...rest} class={btnClass}>
 	<Menu class={menuClass} />
 </ToolbarButton>
 
-<!--
+<!--d
 @component
 [Go to docs](https://flowbite-svelte.com/)
 ## Props
