@@ -3,14 +3,14 @@
 	import { WalletOutline } from 'flowbite-svelte-icons';
 	import { authStore } from '$lib/stores/auth.store';
 
-	let open = false;
+	let open = $state(false);
 	function shortenText(text: string): string {
 		return text.slice(0, 8) + '...' + text.slice(-6);
 	}
 </script>
 
 <button
-	on:click={() => (open = true)}
+	onclick={() => (open = true)}
 	class="bg-black text-white font-light rounded-full py-4 px-5 text-xs flex flex-row items-center hover:bg-gray-800 active:bg-gray-700 focus:outline-none focus:ring focus:ring-zinc-400"
 >
 	{shortenText($authStore.principal.toText())}
