@@ -2,30 +2,37 @@
 	import { twMerge } from 'tailwind-merge';
 	import { Button, Dropdown, DropdownItem } from 'flowbite-svelte';
 	import { ChevronDownOutline } from 'flowbite-svelte-icons';
-	import { createEventDispatcher } from 'svelte';
 
-	const dispatch = createEventDispatcher();
-
-	export let value: string;
-	export let placeholder: string = 'Select option...';
-	export let dropDownClass: string = '';
+	let {
+		items,
+		value = $bindable(),
+		placeholder = 'Select option...',
+		dropDownClass = '',
+		class: className,
+		change = () => {}
+	}: {
+		items: { id: number; value: string; img: string; name: string }[];
+		value: string;
+		placeholder: string;
+		dropDownClass?: string;
+		class?: string;
+		change?: (value: string, selectImg: string) => void;
+	} = $props();
 
 	let activeClass = 'font-medium font-bold text-center py-2 px-4 text-sm bg-slate-300';
 	let defaultClass =
 		'font-medium font-bold text-center py-2 px-4 text-sm hover:bg-gray-100 dark:hover:bg-gray-600';
 
-	let selectImg: string;
-	let selectName: string;
+	let selectImg: string = $state('');
+	let selectName: string = $state('');
 
-	let dropdownOpen = false;
-
-	export let items: { id: number; value: string; img: string; name: string }[];
+	let dropdownOpen = $state(false);
 
 	function handleSelect(item: { id: number; value: string; img: string; name: string }) {
 		value = item.value;
 		selectImg = item.img;
 		selectName = item.name;
-		dispatch('change', { value, selectImg });
+		change(value, selectImg);
 		dropdownOpen = false;
 	}
 
@@ -33,7 +40,7 @@
 		if (item.value === value) {
 			selectImg = item.img;
 			selectName = item.name;
-			dispatch('change', { value, selectImg });
+			change(value, selectImg);
 		}
 	});
 </script>
@@ -41,7 +48,7 @@
 <Button
 	class={twMerge(
 		'text-black w-40 h-16 bg-white hover:bg-slate-300 border-2 drop-shadow-md flex justify-between',
-		$$props.class
+		className
 	)}
 >
 	{#if selectImg}
